+++ source/README.md
@@ -1,14 +1,30 @@
 # CA Web Developer Portfolio
 This portfolio project belongs to Carlos Andres and is a compilation of my best works in web development, showcasing both front-end and back-end skills. The project was developed as part of my studies in the Interactive Media Design program at Fanshawe College.
 
+## Table of Contents
+- [Installation](#installation)
+- [Usage](#usage)
+- [Features](#features)
+- [History](#history)
+- [Credits](#credits)
+- [License](#license)
+
 ## Installation
 1. Clone or download the repository from the provided link;
 2. On the repository page, click the green "< > Code" button;
 3. Select "Download ZIP" and unzip the folder.
 
 ## Usage
-* Download the folder onto your device;
-* Open the index.html file in the browser of your choice to explore the portfolio.
+To explore the portfolio, simply open the index.html file in your browser:
+
+1. Download the folder onto your device.
+2. Open the index.html file in the browser of your choice.
+
+## Features
+- Responsive Design: The portfolio is designed to be fully responsive, ensuring a seamless experience across all devices.
+- Interactive Elements: Includes smooth animations and interactive elements to enhance user engagement.
+- Case Studies: Detailed case studies on various projects, including Billy Beer and others.
+- Contact Form: A functional contact form to get in touch with potential clients and employers.
 
 ## History
 The portfolio was created to display a variety of web development projects, demonstrating my knowledge of HTML, CSS, JavaScript, and back-end integration. The goal is to present these works to potential employers, clients, and industry professionals. It includes interactive elements, smooth animations, and case studies on the projects I’ve worked on, such as Billy Beer and others.
